Norwegian (Bokmaal) translation updated thanks to Arne Hanssen
Bokmaal/Book Language is the dominating language, but is actually
norwegain-danish as Norway was ruled from Denmark for 400 years (the
fourhundred year night).  All governmental duties were carried out by
danes and upper class norwegians strived to speak and write (a kind of)
danish.  And so on - you know, everyone wants to belong to the "better"
class.  Probably 85% of us use mostly bokmaal, including me.

In the late 18-hundreds one man (Ivar Aasen) wanted to preserve the
norwegian language throughout the country and did an impressiv attempt
to cover most of the country.  This is now called Nynorsk/New
Norwegian.  Nynorsk has a very low status in large parts of our country,
even in my part of Norway (Nothern Norway), in spite of our spoken
dialect which is closer to Nynorsk than Bokmaal.  I must admit that am
somewhat ashamed that I'm not quite fluent in Nynorsk, even though the
difference isn't big.  Nynorsk is a "bigger" language in terms of number
of words in its dictionary (about 50% more words than Bokmaal).

In simple terms:  Bokmaal = Danish, Nynorsk = Norwegian (mostly used in
Western Norway).

For the record, Nynorsk is quite different from the "old norwegian" used
by the vikings for 1000 years or so ago.  Icelandic is very much like
"old norwegian".



git-svn-id: https://svn.code.sf.net/p/beebeep/code@1238 127b15f6-8070-407b-945c-913c06b37495